+++ docs/changelog.rst
@@ -27,6 +27,7 @@ fixes
 - Many Fixed TYPOS and grammatical language error
 - Fix number of line when print runfile error
 - Fix flush console message stream when wait process
+- Fix particular limits for var D and I
 
 
 0.4.2 (**2012-08-21**)